+++ apps/web/app/(navigation)/(singleCard)/MainPageTranslated.tsx
@@ -2,8 +2,6 @@
 
 import { TypoStyled } from "components/HomePageComponents";
 import { useTranslate } from "lib/hooks";
-import { signIn, signOut, useSession } from "next-auth/react";
-import { Button } from "ui";
 import styled from "styled-components";
 
 const ContentWrapperStyled = styled.div`
